Revised directives, added many tests
A directive is now clear in the AST as, e.g., (directive (show)).
This comes at the cost of allowing spaces between the '#' and the rest of the directive.
However, I see no reason why this would lead to an issue.
And, I think this is preferable to either omitting the directive type, or including the '#' in the directive type and leaving the directive expression as helper syntax.